Move JSON codecs from wrapper to common internal
