+++ shell/platform/common/cpp/client_wrapper/core_wrapper_files.gni
@@ -8,9 +8,6 @@ core_cpp_client_wrapper_includes =
                     "include/flutter/binary_messenger.h",
                     "include/flutter/encodable_value.h",
                     "include/flutter/engine_method_result.h",
-                    "include/flutter/json_message_codec.h",
-                    "include/flutter/json_method_codec.h",
-                    "include/flutter/json_type.h",
                     "include/flutter/message_codec.h",
                     "include/flutter/method_call.h",
                     "include/flutter/method_channel.h",
@@ -31,8 +28,6 @@ core_cpp_client_wrapper_sources =
         [
           "byte_stream_wrappers.h",
           "engine_method_result.cc",
-          "json_message_codec.cc",  # TODO combine into a single json_codec.cc.
-          "json_method_codec.cc",  # TODO combine into a single json_codec.cc.
           "plugin_registrar.cc",
           "standard_codec_serializer.h",
           "standard_codec.cc",
