+++ shell/platform/common/cpp/json_message_codec.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "include/flutter/json_message_codec.h"
+#include "flutter/shell/platform/common/cpp/json_message_codec.h"
 
 #include <iostream>
 #include <string>
