+++ shell/platform/common/cpp/json_message_codec.h
