+++ shell/platform/glfw/key_event_handler.cc
@@ -6,7 +6,7 @@
 
 #include <iostream>
 
-#include "flutter/shell/platform/common/cpp/client_wrapper/include/flutter/json_message_codec.h"
+#include "flutter/shell/platform/common/cpp/json_message_codec.h"
 
 static constexpr char kChannelName[] = "flutter/keyevent";
