+++ shell/platform/windows/key_event_handler.cc
@@ -8,7 +8,7 @@
 
 #include <iostream>
 
-#include "flutter/shell/platform/common/cpp/client_wrapper/include/flutter/json_message_codec.h"
+#include "flutter/shell/platform/common/cpp/json_message_codec.h"
 
 static constexpr char kChannelName[] = "flutter/keyevent";
